🔧 Fix: Fix issue with retrieving unique indexes in MySQL wrapper
The MySQL wrapper was not correctly identifying unique indexes due to a
error in the query. The `NON_UNIQUE` field was used to determine uniqueness,
but it was negated using the `NOT` operator. This resulted in incorrect
values for the `unique` property in the returned indexes. The query has
been updated to use the `NON_UNIQUE` field directly without negating it,
fixing the issue with identifying unique indexes.

This change fixes the bug in the MySQL wrapper and ensures that the unique
property of indexes is correctly determined. It improves the accuracy of
index information when working with MySQL databases.