feat(readme): Add MySQL and Monaco updates
Added MySQL and Monaco updates to the README.md file. Updated the CRUD operations with the float, decimal, and "bigNumber" values. Also, included the use of tunnel in the bash command. For Monaco, added autocomplete functionality with the use of crypto-js library and added a feature to reload the app when the structure changes.

refactor(sql): Refactor delete and update functions

Refactored the delete and update functions in the SQL class. Added a new method called `getPk()` to get the primary key(s) from the indexes. The `delete()` function now uses the primary key(s) to build the `WHERE` clause for deletion. Similarly, the `update()` function now gets the primary key(s), compares old and new data, and builds the update query accordingly.

fix(container): Update container component HTML template

Updated the container component's HTML template. Added a tooltip to display server details including wrapper, user, and params. Also, added a button to reload the database. Removed the previous reload button from the context menu.

feat(container): Add displayParams() method

Added a new method called `displayParams()` in the `ContainerComponent` class. This method returns the server params object as a JSON string to be displayed in the tooltip. This helps in viewing the server params in a more readable format.