bug: Fix issue with MySQL index query
The MySQL query used to retrieve indexes was incorrect. The `NOT NON_UNIQUE`
part of the query was causing incorrect data to be returned. This has been
fixed by removing `NOT NON_UNIQUE` from the query.

:art: Refactor PostgreSQL index query

The PostgreSQL query used to retrieve indexes has been refactored for
improved performance and readability. The query now includes additional
information such as the cardinality of the indexes. The indexes are also
now grouped and returned in a more structured format.

:sparkles: Add custom route reuse strategy

A custom route reuse strategy has been implemented to cache certain routes
in Angular. The strategy caches routes with specific paths, such as
'explore' and 'query', to improve performance by reusing the component
instances instead of creating new ones.

:sparkles: Add custom reuse strategy module

A new module `custom-reuse-strategy.ts` has been added to handle the
custom route reuse strategy. This module contains the implementation of
the `CustomReuseStrategy` class which extends the `RouteReuseStrategy`
interface and defines the custom behavior for route reuse.

:recycle: Refactor PostgreSQL driver

The PostgreSQL driver class has been refactored to improve the code
structure and readability. The functions and constants have been
reorganized for better organization and maintainability. The driver now
includes a `format` method to format SQL code using the `sql-formatter`
library.

:art: Improve code formatting in code component

The code formatting in the code component has been improved to include
line breaks after each line of code. This improves the readability of the
code and makes it easier to distinguish between different lines.