✨ `:fire:` Remove unused dependency from front/pnpm-lock.yaml
The `sql-highlight` dependency was removed from the `front/pnpm-lock.yaml` file. This commit removes the unnecessary code.

The dependency was removed because it was no longer being used in the project and was causing unnecessary bloating of the codebase. Removing unused dependencies helps to improve the maintainability and efficiency of the project.