🐛 Use `$angular-accent` variable instead of hardcoded color
The color of the `.toUpdate` class and its child `.mat-select-value` is changed from `#ff9800` to `mat-color($angular-accent, default)` to make it more consistent with the Angular color palette. The change ensures that the color can be easily updated globally by modifying the `$angular-accent` variable.

These changes were made to enhance the maintainability of the codebase and adhere to consistent design standards.