art: Refactor(front): Rename label in explore.component.html for paginator
This commit updates the label in the explore.component.html file for the
mat-form-field used in the paginator section. The label "Go To Page" has been
renamed to "Page" for better clarity and simplicity.

:bug: Fix(front): Use const instead of let in explore.component.ts

This commit fixes a bug in the explore.component.ts file where the 'let' keyword
was used instead of the 'const' keyword. The affected variable 'params' is now
declared as a constant using the 'const' keyword.

:art: Refactor(front): Rename label in code.component.html for paginator

This commit updates the label in the code.component.html file for the
mat-form-field used in the paginator section. The label "Go To Page" has been
renamed to "Page" for better clarity and simplicity.