📝 Added support for autocomplete table and column names and types.
The patch includes changes to several files:
- Modified `README.md`
- Updated `structure.component.html`
- Updated `mongodb.ts`
- Updated `generator.ts`
- Updated `code.component.html`
- Updated `code.module.ts`

In the `structure.component.html` and `code.component.html` files, the `ngx-json-viewer` component was modified to have the `expanded` property set to `false` when displaying JSON data.

In the `mongodb.ts` file, the `language` property of the `MongoDB` class was changed from `'javascript'` to `'typescript'`.

In the `generator.ts` file, a typo was fixed in a function call.

In the `code.module.ts` file, unnecessary code related to dynamic loading of driver languages was removed, and autocomplete suggestions were added for TypeScript and SQL language. Additionally, an extra library was added for the `crypto-js` package to provide autocomplete suggestions for specific functions.

The changes were made to enhance the user experience by providing autocomplete suggestions and improving the display of JSON data.