card_file_box: (back) Update `dev` script in `package.json`
This commit updates the `dev` script in the `package.json` file for the backend. The script is renamed to `dev:back` to improve clarity and distinguish it from other scripts. The purpose of this change is to make it easier for developers to run and develop the backend server.

:hammer_and_wrench: (back) Refactor `mongodb.js` in `src/wrapper`

This commit refactors the `mongodb.js` file in the `src/wrapper` directory of the backend.

- Imports the `MongoClient` and `ObjectId` from the `"mongodb"` package.
- Removes unused imports and commented code.
- Updates the import statement for the `Driver` class.

The purpose of this change is to optimize the import statements and enhance the code readability in the `mongodb.js` file.

:card_file_box: (back) Update `runCommand` method in `mongodb.js`

This commit updates the `runCommand` method in the `mongodb.js` file of the backend.

- Modifies the parameter of the method to accept `database` as an optional parameter.
- Updates the implementation of the method to use the `evalInContext` function instead of the `getResponse` function.
- Uses the `db` object from the `connection` instead of accessing the `spawnedShell` directly.
- Returns the result of the execution as an array by calling `result.toArray()`.

The purpose of this change is to improve the execution of commands in the MongoDB shell and provide a more consistent and reliable response.

:card_file_box: (front) Update `dev` script in `package.json`

This commit updates the `dev` script in the `package.json` file for the frontend. The script is renamed to `dev:front` to improve clarity and distinguish it from other scripts. The purpose of this change is to make it easier for developers to run and develop the frontend application.

:building_construction: (front) Update import statement in `container.component.ts`

This commit updates the import statement in the `container.component.ts` file of the frontend.

- Removes the unused import `MatDialogRef` from `"@angular/material/dialog"`.

The purpose of this change is to remove the unused import and improve the code maintainability.

:lipstick: (front) Update button label in `servers.component.html`

This commit updates the button label in the `servers.component.html` file of the frontend.

- Changes the label from "Refresh Connections" to "Rescan Connections" to better reflect the action performed by the button.

The purpose of this change is to improve the clarity and accuracy of the button label for users.