hammer: Refactor MongoDB and SQL drivers
Refactor the `getBaseSelect` and `getBaseSelectWithRelations` methods in the `MongoDB` and `SQL` driver classes. Previously, the `WHERE 1 = 1` condition was added to the SQL queries, but it is not necessary. Removed the condition from both methods to simplify the queries.

:recycle: Rearrange code for code formatting

In the `CodeComponent` class, rearranged the code to trigger the `formatDocument` action on all editors after a brief delay. This ensures that the code is properly formatted for better readability.