fire: Delete custom-reuse-strategy.ts
The file `custom-reuse-strategy.ts` is no longer needed and has been removed. This file was responsible for implementing a custom route reuse strategy in the application. It was decided that this functionality was unnecessary and could be removed to simplify the codebase. The commit removes the file and all references to it.

:art: Improve code in app-routing.module.ts

There were unnecessary code in the `AppRoutingModule` module which have been removed. The code was related to a custom route reuse strategy that is no longer being used in the application. The commit removes the related code to improve the codebase and remove unused dependencies.

:bug: Fix an issue in cell.component.ts

There was a bug in the `getFkParams()` method of `CellComponent`. The bug occurred when constructing the `chips` parameter to be passed to another component. There was an incorrect use of quotes, causing the generated parameter to be invalid. This commit fixes the bug by using the correct type of quotes in the `chips` parameter.

:art: Improve code in code.component.ts

There were some improvements made to the `CodeComponent` code. The initial setup of the component was being done in the constructor, which is not recommended. The setup code has been moved to the `ngOnInit` method. This ensures that the component is fully initialized before any setup is performed. Additionally, the code has been reorganized for better readability.

:recycle: Refactor core-routing.module.ts

The code in the `CoreRoutingModule` module has been refactored for better readability and maintainability. The `data` object for each route now uses the `reuseRouteId` property instead of the deprecated `reuseRoute` property. This change was made to align with the latest conventions and to avoid any potential conflicts or issues with the deprecated property.