✨ Add MongoDB library to front-end
This commit adds the MongoDB library to the front-end code. The `loadExtraLib` function in the MongoDB driver is used to load the `mongo.d.ts` and `bson.d.ts` files from the `assets/libs/` directory. These files are then added as extra libraries to the Monaco Editor, providing additional typings and autocomplete suggestions for MongoDB queries. This improves the developer experience when working with MongoDB queries in the application.