sparkles: Add .gitignore for src/assets/libs
The .gitignore file in the front directory is updated to include the src/assets/libs/ directory. This means that any files or directories within src/assets/libs/ will be ignored by Git.

This change is made to prevent versioning of the assets/libs/ directory, as it might contain large files or external dependencies that don't need to be tracked by Git. Ignoring this directory helps keep the repository clean and avoids unnecessary bloat.

By excluding the src/assets/libs/ directory from version control, it reduces the repository's size and improves performance during operations such as cloning or merging.

The use of a .gitignore file follows best practices for managing version control and aids in managing project dependencies effectively.