feat: 🎉 Add custom reuse strategy for route handling
Add a custom reuse strategy (`CustomReuseStrategy`) to handle route
reuse in the Angular application. The strategy is implemented to
store and retrieve route handles based on a provided `reuseRouteId`
in the route's data.

This change was done to optimize the application's performance by
reusing route components whenever possible, rather than recreating
them each time.

The custom reuse strategy is added to the `AppRoutingModule`
providers, so that it is used throughout the application.

The `core-routing.module.ts` file is also updated to include the
`reuseRouteId` data property for each route component that should be
reuseable.