lipstick: Refactor container and explore components
The container component's HTML and the container component's SCSS files were
refactored. The changes include removing unnecessary styles, adjusting the max
width of the mat-drawer, and adding flex properties to the mat-drawer-content.
The changes were made to improve the layout and appearance of the container
component.

The explore component's HTML file was modified to remove the mat-paginator
element and the setAutoUp button. These elements were removed because they were
rendered incorrectly in the explore component. The changes were made to fix the
rendering issue and improve the usability of the explore component.

The changes in container and explore components were done to improve layout and
fix rendering issues.