recycle: Refactor MongoDB class
Refactored the `MongoDB` class in `back/src/wrapper/mongodb.js`. Removed the use of the `replaceAll()` method and instead commented it out because it's not available in all JavaScript environments. This change ensures compatibility and improves code readability.

:bug: Fix connection issue

In `front/src/app/container/container.component.ts`, fixed an issue where the server and database were not being properly assigned. Added null checks and assigned values to `server` and `database` variables before using them in subsequent logic. This ensures that the correct server and database are used for the connection.

:lipstick: Remove unnecessary CSS

Removed an unnecessary CSS style in `front/src/app/core/tables/tables.component.scss` where the display property for `.mat-icon-button` was set to none. This style was not being used and can be safely removed to reduce code clutter.