art: Remove unnecessary lines in README.md
The README.md file had some unnecessary lines that needed to be removed.
These lines were reviewed and deemed unnecessary for the project.

:arrow_up: Update app-routing.module.ts

In the app-routing.module.ts file, the path for the component was updated.
The previous path included a tab parameter, which was removed in this update.
This change was made to simplify the routing structure of the application.

:recycle: Refactor servers.component.ts

In the servers.component.ts file, some imports were removed and the code
was refactored to improve readability and remove unused dependencies.
This change was made to clean up the code and improve overall performance.

:building_construction: Update core-routing.module.ts

The core-routing.module.ts file was updated to add a default redirect
path when no specific route is matched. It now redirects to the 'explore'
path. This change was made to ensure a smoother user experience and
provide a default landing page when no specific path is specified.

:twisted_rightwards_arrows: Update tables.component.ts

The tables.component.ts file was updated to change the route when changing
the selected table. The URL is updated to remove the current table name
and replace it with the new selected table name. This change was made to
provide a more intuitive and consistent user experience when navigating
between tables.