memo: Fix querySize method in mongodb wrapper
The `querySize` method in the `MongoDB` class was updated to properly count the number of documents returned by a query. The previous implementation did not handle queries with the `.toArray()` method correctly. This commit replaces the `.find()` method with `.countDocuments()` and removes the `.toArray()` method from the query before executing it. This ensures that the correct number of documents is returned when querying the database.

:recycle: Refactor explore.component.html

In the `explore.component.html` file, the `*cdkVirtualFor` directive was replaced with `*ngFor` to iterate over the `displayedColumns` array. This change was made to align with the Angular template syntax. Additionally, the `cdk-virtual-scroll-viewport` element was replaced with a `div` element with the class "scrollDiv" in order to fix the styling of the virtual scroll container.

:bug: Fix URL replacement in changeTable method

The `changeTable` method in the `TablesComponent` class had an issue with replacing the table name in the URL. The previous implementation used the `replace` method with a hardcoded separator ("/") and assumed that the selected table name would always be surrounded by slashes. This commit updates the URL replacement logic to use template literals and replaces the selected table name without making any assumptions about the URL structure. This ensures that the table name is replaced correctly in the URL when changing tables.